testing submodule and small macro change
